Rename files folder to data
