+++ backend/apps/matching/data/make_plz_json_script.py
@@ -13,7 +13,7 @@ current_location = dirname(abspath(__file__))
 from os.path import join
 
 for countrycode in ["DE", "AT"]:
-    with open(join(current_location, f'files/{countrycode}.csv'), encoding='utf-8') as csvfile:
+    with open(join(current_location, f'data/{countrycode}.csv'), encoding='utf-8') as csvfile:
         reader = csv.DictReader(csvfile)
         plzs[countrycode] = {}
         for row in reader:
@@ -30,6 +30,6 @@ for countrycode in ["DE", "AT"]:
 
 
 
-with open(join(current_location, "files/zipcodes_merged.json"), "w") as f:
+with open(join(current_location, "data/zipcodes_merged.json"), "w") as f:
     f.write(json.dumps(plzs))
 """
