+++ backend/apps/matching/data/map_data.py
@@ -3,5 +3,5 @@ from os.path import abspath, dirname, join
 
 current_location = dirname(abspath(__file__))
 
-with open(join(current_location, "../files/zipcodes_merged.json")) as f:
+with open(join(current_location, "../data/zipcodes_merged.json")) as f:
     zipcodes = json.loads(f.read())
