+++ backend/apps/matching/models/participant_info_location.py
@@ -6,7 +6,7 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 import numpy as np
 
-from apps.matching.files.map_data import zipcodes
+from apps.matching.data.map_data import zipcodes
 from apps.matching.utils.zipcodes import GERMAN_BIG_CITY_ZIPCODES
 
 from .participant_info import ParticipantInfo
